Add config.hops to change peering behavior
Problem: Peering with only people that we explicitly follow can cause
problems when the people you follow don't have public IP addresses.
Since this isn't configurable, interface developers have to resort to
hacks that auto-connect to staged peers, which introduce lots of subtle
bugs throughout the stack.

Solution: Make the peering behavior configurable, with a default where
we only connect to peers that are we explicitly follow.